Split oversized admin panels into hooks and subcomponents
Contacts, families, and organisations:
- Extract panel editor hooks and editor/list subcomponents
- Add shared entity-remove-member dialog

Service and instance detail:
- Move form state and handlers into dedicated hooks
- Extract service-detail-form-body and instance type sections
- Extract instance payload builders and form-state helpers

Contacts editor:
- Split referral search and service-label loading into focused hooks
  so use-contacts-panel-editor stays under 500 lines

All six original panel files are now under 500 lines; supporting hooks
and subcomponents follow the same limit.
