+++ apps/admin_web/src/components/admin/contacts/contacts-panel.tsx
@@ -1,68 +1,15 @@
 'use client';
 
-import { useEffect, useMemo, useState, type MouseEvent } from 'react';
-
 import type { useAdminEntityContacts } from '@/hooks/use-admin-entity-contacts';
-import { useFamilyOrgPickers } from '@/hooks/use-family-org-pickers';
-import { useGeocodeVenueAddress } from '@/hooks/use-geocode-venue-address';
-import { useInlineLocationSave } from '@/hooks/use-inline-location-save';
-import { useConfirmDialog } from '@/hooks/use-confirm-dialog';
-import { InlineLocationEditor } from '@/components/admin/locations/inline-location-editor';
-import type { InlineLocationEmbeddedSummary } from '@/components/admin/locations/inline-location-editor';
+import { useContactsPanelEditor } from '@/hooks/use-contacts-panel-editor';
+import { ContactEditorCard } from '@/components/admin/contacts/contact-editor-card';
 import { ContactNotesPanel } from '@/components/admin/contacts/contact-notes-panel';
-import { ContactEditorIdentityFields } from '@/components/admin/contacts/contact-editor-identity-fields';
-import { ContactEditorSourceFields } from '@/components/admin/contacts/contact-editor-source-fields';
+import { ContactsListTable } from '@/components/admin/contacts/contacts-list-table';
 import { MailchimpSyncCard } from '@/components/admin/contacts/mailchimp-sync-card';
-import { EntityServicesSection } from '@/components/admin/contacts/entity-services-section';
-import { EntityTagPicker } from '@/components/admin/contacts/entity-tag-picker';
-import { ArchiveIcon, DeleteIcon, NoteIcon, RestoreIcon } from '@/components/icons/action-icons';
-import { Button } from '@/components/ui/button';
 import { ConfirmDialog } from '@/components/ui/confirm-dialog';
-import { AdminCollapsibleSection } from '@/components/ui/admin-collapsible-section';
-import {
-  AdminDataTable,
-  AdminDataTableBody,
-  AdminDataTableCell,
-  AdminDataTableHead,
-  AdminDataTableHeadCell,
-  AdminDataTableOperationsHeadCell,
-} from '@/components/ui/admin-data-table';
-import { AdminEditorCard } from '@/components/ui/admin-editor-card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { PaginatedTableCard } from '@/components/ui/paginated-table-card';
-import { AdminTableToolbar } from '@/components/ui/admin-table-toolbar';
-import { Select } from '@/components/ui/select';
-import {
-  getAdminContact,
-  listAdminContactServices,
-  searchEntityContactsForPicker,
-  type EntityTagRef,
-} from '@/lib/entity-api';
-import {
-  formatAdminContactPickerLabel,
-  formatEntityVenueLocationLabel,
-  formatEnumLabel,
-} from '@/lib/format';
-import { contactPhoneRequestFields } from '@/lib/phone-request';
-import { CONTACT_TYPES } from '@/lib/contacts/contacts-panel-constants';
-import type { EntityPickerListItem } from '@/lib/entity-api';
-import type { EntityListFilters } from '@/types/entity-list';
-import {
-  CONTACT_RELATIONSHIP_TYPES,
-  relationshipTypeForEditor,
-} from '@/types/entity-relationship';
+import type { EntityTagRef } from '@/lib/entity-api';
 import type { AdminUser } from '@/types/leads';
 import type { GeographicAreaSummary, LocationSummary } from '@/types/services';
-import type { components } from '@/types/generated/admin-api.generated';
-
-import {
-  contactNameListSuffix,
-  contactRowLabel,
-  linkedVenueReadOnlyLines,
-} from '@/lib/contacts/contacts-panel-helpers';
-
-type ApiSchemas = components['schemas'];
 
 export interface ContactsPanelProps {
   contacts: ReturnType<typeof useAdminEntityContacts>;
@@ -73,7 +20,6 @@ export interface ContactsPanelProps {
   geographicAreas: GeographicAreaSummary[];
   areasLoading: boolean;
   refreshLocations: () => Promise<void> | void;
-  /** Refetch family and organisation lists (membership can change when a contact is saved). */
   refreshFamilyOrgLists?: () => void | Promise<void>;
 }
 
@@ -97,765 +43,59 @@ export function ContactsPanel({
     hasMore,
     error,
     loadMore,
-    isSaving,
-    createContact,
-    updateContact,
-    deleteContact,
   } = contacts;
 
-  const [confirmDialogProps, requestConfirm] = useConfirmDialog();
-  const [pendingLocationLeaveDialogProps, requestPendingLocationLeaveConfirm] = useConfirmDialog();
-  const [deleteActionError, setDeleteActionError] = useState('');
-  const [notesTarget, setNotesTarget] = useState<ApiSchemas['AdminContact'] | null>(null);
-  const { familyPicker, organizationPicker } = useFamilyOrgPickers();
-
-  const [editorMode, setEditorMode] = useState<'create' | 'edit'>('create');
-  const [selectedId, setSelectedId] = useState<string | null>(null);
-  const [firstName, setFirstName] = useState('');
-  const [lastName, setLastName] = useState('');
-  const [email, setEmail] = useState('');
-  const [instagramHandle, setInstagramHandle] = useState('');
-  const [phoneRegion, setPhoneRegion] = useState('HK');
-  const [phoneNational, setPhoneNational] = useState('');
-  const [contactType, setContactType] = useState<ApiSchemas['EntityContactType']>('parent');
-  const [relationshipType, setRelationshipType] =
-    useState<(typeof CONTACT_RELATIONSHIP_TYPES)[number]>('prospect');
-  const [source, setSource] = useState<ApiSchemas['EntityContactSource']>('manual');
-  const [sourceDetail, setSourceDetail] = useState('');
-  const [referralContactId, setReferralContactId] = useState('');
-  const [referralSearchInput, setReferralSearchInput] = useState('');
-  const [referralSearchResults, setReferralSearchResults] = useState<EntityPickerListItem[]>([]);
-  const [referralPinnedLabel, setReferralPinnedLabel] = useState('');
-  const [dateOfBirth, setDateOfBirth] = useState('');
-  const [pendingLocationId, setPendingLocationId] = useState<string | null>(null);
-  const [optimisticLocationSummary, setOptimisticLocationSummary] =
-    useState<InlineLocationEmbeddedSummary | null>(null);
-  const [familySelectId, setFamilySelectId] = useState('');
-  const [organizationSelectId, setOrganizationSelectId] = useState('');
-  const [tagIds, setTagIds] = useState<string[]>([]);
-  const [serviceLabelsState, setServiceLabelsState] = useState<{
-    entityId: string;
-    labels: string[];
-  } | null>(null);
-  const [active, setActive] = useState(true);
-
-  const serviceLabels =
-    editorMode === 'edit' && selectedId && serviceLabelsState?.entityId === selectedId
-      ? serviceLabelsState.labels
-      : [];
-
-  const {
-    status: locationSaveStatus,
-    createSharedLocation,
-    updateSharedLocation,
-    clearError: clearLocationSaveError,
-  } = useInlineLocationSave(refreshLocations);
-  const { geocode: geocodeLocation, isGeocoding: locationGeocoding } = useGeocodeVenueAddress();
-
-  const selected = useMemo(
-    () => rows.find((c) => c.id === selectedId) ?? null,
-    [rows, selectedId]
-  );
-
-  const linkedToFamilyOrOrg = Boolean(familySelectId.trim() || organizationSelectId.trim());
-  const locationFieldLocked = linkedToFamilyOrOrg;
-
-  const readOnlyLockedLinesForEditor = useMemo(() => {
-    if (!linkedToFamilyOrOrg || !selected) {
-      return null;
-    }
-    const { lines, footerNote } = linkedVenueReadOnlyLines(selected);
-    if (lines.length === 0) {
-      return null;
-    }
-    return { lines, footerNote };
-  }, [linkedToFamilyOrOrg, selected]);
-
-  const inlineLocationStateKey =
-    editorMode === 'create' ? 'contact-new' : `contact:${selectedId ?? 'none'}`;
-
-  const resolvedLocation = useMemo(() => {
-    if (!pendingLocationId) {
-      return null;
-    }
-    return locations.find((l) => l.id === pendingLocationId) ?? null;
-  }, [locations, pendingLocationId]);
-
-  const embeddedLocationSummary = useMemo((): InlineLocationEmbeddedSummary | null => {
-    if (resolvedLocation) {
-      return null;
-    }
-    if (!pendingLocationId) {
-      return null;
-    }
-    if (optimisticLocationSummary && optimisticLocationSummary.id === pendingLocationId) {
-      return optimisticLocationSummary;
-    }
-    const s = selected?.location_summary;
-    if (s && s.id === pendingLocationId) {
-      return {
-        id: s.id,
-        name: s.name ?? null,
-        address: s.address ?? null,
-        areaName: s.area_name,
-        areaId: s.area_id,
-        lat: s.lat ?? null,
-        lng: s.lng ?? null,
-      };
-    }
-    return null;
-  }, [resolvedLocation, pendingLocationId, optimisticLocationSummary, selected?.location_summary]);
-
-  function summaryFromLocationRow(loc: LocationSummary): InlineLocationEmbeddedSummary {
-    const areaName = geographicAreas.find((a) => a.id === loc.areaId)?.name ?? 'Unknown area';
-    return {
-      id: loc.id,
-      name: loc.name,
-      address: loc.address,
-      areaName,
-      areaId: loc.areaId,
-      lat: loc.lat,
-      lng: loc.lng,
-    };
-  }
-
-  const referralSelectOptions = useMemo(() => {
-    const byId = new Map<string, string>();
-    for (const r of referralSearchResults) {
-      byId.set(r.id, r.label);
-    }
-    const rid = referralContactId.trim();
-    if (rid && referralPinnedLabel.trim() && !byId.has(rid)) {
-      byId.set(rid, referralPinnedLabel.trim());
-    }
-    return Array.from(byId.entries()).map(([id, label]) => ({ id, label }));
-  }, [referralSearchResults, referralContactId, referralPinnedLabel]);
-
-  useEffect(() => {
-    if (source !== 'referral') {
-      queueMicrotask(() => {
-        setReferralSearchResults([]);
-      });
-      return;
-    }
-    const q = referralSearchInput.trim();
-    if (q.length < 2) {
-      queueMicrotask(() => {
-        setReferralSearchResults([]);
-      });
-      return;
-    }
-    let cancelled = false;
-    const handle = setTimeout(() => {
-      void (async () => {
-        try {
-          const items = await searchEntityContactsForPicker({
-            query: q,
-            excludeContactId: editorMode === 'edit' ? selectedId : null,
-            limit: 50,
-          });
-          if (!cancelled) {
-            setReferralSearchResults(Array.isArray(items) ? items : []);
-          }
-        } catch {
-          if (!cancelled) {
-            setReferralSearchResults([]);
-          }
-        }
-      })();
-    }, 300);
-    return () => {
-      cancelled = true;
-      clearTimeout(handle);
-    };
-  }, [referralSearchInput, source, selectedId, editorMode]);
-
-  useEffect(() => {
-    if (source !== 'referral' || !referralContactId.trim()) {
-      queueMicrotask(() => {
-        setReferralPinnedLabel('');
-      });
-      return;
-    }
-    const id = referralContactId.trim();
-    let cancelled = false;
-    void (async () => {
-      try {
-        const c = await getAdminContact(id);
-        if (cancelled || !c) {
-          return;
-        }
-        setReferralPinnedLabel(formatAdminContactPickerLabel(c));
-      } catch {
-        if (!cancelled) {
-          setReferralPinnedLabel('');
-        }
-      }
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, [referralContactId, source]);
-
-  useEffect(() => {
-    if (editorMode !== 'edit' || !selectedId) {
-      return;
-    }
-    const entityId = selectedId;
-    const controller = new AbortController();
-    let cancelled = false;
-    void (async () => {
-      try {
-        const labels = await listAdminContactServices(entityId, controller.signal);
-        if (!cancelled) {
-          setServiceLabelsState({ entityId, labels });
-        }
-      } catch {
-        if (!cancelled) {
-          setServiceLabelsState({ entityId, labels: [] });
-        }
-      }
-    })();
-    return () => {
-      cancelled = true;
-      controller.abort();
-    };
-  }, [selectedId, editorMode]);
-
-  async function resetCreateForm() {
-    if (editorMode === 'create' && pendingLocationId) {
-      const ok = await requestPendingLocationLeaveConfirm({
-        title: 'Leave without finishing?',
-        description:
-          'You saved an address to a new location but have not finished creating this contact yet. Leave anyway? The location row stays in the directory.',
-        confirmLabel: 'Leave',
-        cancelLabel: 'Stay',
-        variant: 'default',
-      });
-      if (!ok) {
-        return;
-      }
-    }
-    setEditorMode('create');
-    setSelectedId(null);
-    setFirstName('');
-    setLastName('');
-    setEmail('');
-    setInstagramHandle('');
-    setPhoneRegion('HK');
-    setPhoneNational('');
-    setContactType('parent');
-    setRelationshipType('prospect');
-    setSource('manual');
-    setSourceDetail('');
-    setReferralContactId('');
-    setReferralSearchInput('');
-    setReferralSearchResults([]);
-    setReferralPinnedLabel('');
-    setDateOfBirth('');
-    setPendingLocationId(null);
-    setOptimisticLocationSummary(null);
-    clearLocationSaveError();
-    setFamilySelectId('');
-    setOrganizationSelectId('');
-    setTagIds([]);
-    setActive(true);
-  }
-
-  async function handleSubmit(): Promise<void> {
-    try {
-      const dob = dateOfBirth.trim() ? dateOfBirth.trim() : null;
-      const loc = pendingLocationId;
-      const fam = familySelectId.trim();
-      const org = organizationSelectId.trim();
-      const family_ids = fam ? [fam] : [];
-      const organization_ids = org ? [org] : [];
-
-      if (source === 'referral' && !referralContactId.trim()) {
-        return;
-      }
-
-      if (editorMode === 'create') {
-        await createContact({
-          first_name: firstName.trim(),
-          last_name: lastName.trim() || null,
-          email: email.trim() || null,
-          instagram_handle: instagramHandle.trim() || null,
-          ...contactPhoneRequestFields(phoneRegion, phoneNational),
-          contact_type: contactType,
-          relationship_type: relationshipType,
-          source,
-          source_detail: sourceDetail.trim() || null,
-          date_of_birth: dob,
-          location_id: linkedToFamilyOrOrg ? null : loc,
-          tag_ids: tagIds,
-          family_ids,
-          organization_ids,
-          referral_contact_id:
-            source === 'referral' ? referralContactId.trim() : null,
-        });
-        await refreshFamilyOrgLists?.();
-        await resetCreateForm();
-        return;
-      }
-      if (!selected) {
-        return;
-      }
-      const body: ApiSchemas['UpdateAdminContactRequest'] = {
-        first_name: firstName.trim(),
-        last_name: lastName.trim() || null,
-        email: email.trim() || null,
-        instagram_handle: instagramHandle.trim() || null,
-        ...contactPhoneRequestFields(phoneRegion, phoneNational),
-        contact_type: contactType,
-        relationship_type: relationshipType,
-        source,
-        source_detail: sourceDetail.trim() || null,
-        date_of_birth: dob,
-        active,
-        tag_ids: tagIds,
-        family_ids,
-        organization_ids,
-      };
-      if (source === 'referral') {
-        body.referral_contact_id = referralContactId.trim();
-      }
-      if (!locationFieldLocked) {
-        body.location_id = loc;
-      }
-      await updateContact(selected.id, body);
-      await refreshFamilyOrgLists?.();
-    } catch {
-      // Retry with form state preserved.
-    }
-  }
-
-  async function handleDeleteContact(
-    row: ApiSchemas['AdminContact'],
-    clickEvent: MouseEvent<HTMLButtonElement>
-  ): Promise<void> {
-    clickEvent.stopPropagation();
-    const confirmed = await requestConfirm({
-      title: 'Delete contact',
-      description: `Permanently delete "${contactRowLabel(row)}"? This removes the contact from the database and cannot be undone.`,
-      confirmLabel: 'Delete',
-      cancelLabel: 'Cancel',
-      variant: 'danger',
-    });
-    if (!confirmed) {
-      return;
-    }
-    setDeleteActionError('');
-    try {
-      await deleteContact(row.id);
-      await refreshFamilyOrgLists?.();
-      if (selectedId === row.id) {
-        await resetCreateForm();
-      }
-    } catch (err) {
-      setDeleteActionError(err instanceof Error ? err.message : 'Failed to delete contact');
-    }
-  }
-
-  function selectRow(row: ApiSchemas['AdminContact']) {
-    setSelectedId(row.id);
-    setEditorMode('edit');
-    setFirstName(row.first_name);
-    setLastName(row.last_name ?? '');
-    setEmail(row.email ?? '');
-    setInstagramHandle(row.instagram_handle ?? '');
-    setPhoneRegion(row.phone_region ?? 'HK');
-    setPhoneNational(row.phone_national_number ?? '');
-    setContactType(row.contact_type);
-    setRelationshipType(relationshipTypeForEditor(row.relationship_type));
-    setSource(row.source);
-    setSourceDetail(row.source_detail ?? '');
-    setReferralContactId(row.referral_contact_id ?? '');
-    setReferralSearchInput('');
-    setReferralSearchResults([]);
-    setReferralPinnedLabel('');
-    setDateOfBirth(row.date_of_birth ?? '');
-    setPendingLocationId(row.location_id ?? null);
-    setOptimisticLocationSummary(null);
-    clearLocationSaveError();
-    setFamilySelectId(row.family_ids[0] ?? '');
-    setOrganizationSelectId(row.organization_ids[0] ?? '');
-    setTagIds([...row.tag_ids]);
-    setActive(row.active);
-  }
-
-  const saveDisabled =
-    isSaving ||
-    !firstName.trim() ||
-    (source === 'referral' && !referralContactId.trim());
+  const editor = useContactsPanelEditor({
+    contacts,
+    locations,
+    geographicAreas,
+    refreshLocations,
+    refreshFamilyOrgLists,
+  });
 
   return (
     <div className='space-y-6'>
-      <ConfirmDialog {...confirmDialogProps} />
-      <ConfirmDialog {...pendingLocationLeaveDialogProps} />
+      <ConfirmDialog {...editor.confirmDialogProps} />
+      <ConfirmDialog {...editor.pendingLocationLeaveDialogProps} />
       <MailchimpSyncCard />
-      <AdminEditorCard
-        title='Contact'
-        description='Create a contact or select a row below to edit. When this contact is linked to a family or organisation, set location on that record instead. Mailchimp sync status is read-only from the API.'
-        actions={
-          <>
-            {editorMode === 'edit' ? (
-              <Button
-                type='button'
-                variant='secondary'
-                onClick={() => void resetCreateForm()}
-                disabled={isSaving}
-              >
-                Cancel
-              </Button>
-            ) : null}
-            <Button type='button' disabled={saveDisabled} onClick={() => void handleSubmit()}>
-              {editorMode === 'create' ? 'Create contact' : 'Update contact'}
-            </Button>
-          </>
-        }
-      >
-        <div className='space-y-4'>
-          <ContactEditorIdentityFields
-            firstName={firstName}
-            lastName={lastName}
-            contactType={contactType}
-            relationshipType={relationshipType}
-            email={email}
-            phoneRegion={phoneRegion}
-            phoneNational={phoneNational}
-            instagramHandle={instagramHandle}
-            dateOfBirth={dateOfBirth}
-            onFirstNameChange={setFirstName}
-            onLastNameChange={setLastName}
-            onContactTypeChange={setContactType}
-            onRelationshipTypeChange={setRelationshipType}
-            onEmailChange={setEmail}
-            onPhoneRegionChange={setPhoneRegion}
-            onPhoneNationalChange={setPhoneNational}
-            onInstagramHandleChange={setInstagramHandle}
-            onDateOfBirthChange={setDateOfBirth}
-          />
-
-          <ContactEditorSourceFields
-            source={source}
-            sourceDetail={sourceDetail}
-            referralContactId={referralContactId}
-            referralSearchInput={referralSearchInput}
-            referralSelectOptions={referralSelectOptions}
-            onSourceChange={(v) => {
-              setSource(v);
-              if (v !== 'referral') {
-                setReferralContactId('');
-                setReferralSearchInput('');
-                setReferralSearchResults([]);
-                setReferralPinnedLabel('');
-              } else {
-                setReferralSearchResults([]);
-              }
-            }}
-            onSourceDetailChange={setSourceDetail}
-            onReferralSearchInputChange={setReferralSearchInput}
-            onReferralContactIdChange={(contactId, pinnedLabel) => {
-              setReferralContactId(contactId);
-              if (pinnedLabel) {
-                setReferralPinnedLabel(pinnedLabel);
-              }
-            }}
-          />
-
-          <div className='grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3'>
-            <div>
-              <Label htmlFor='crm-contact-family'>Family</Label>
-              <Select
-                id='crm-contact-family'
-                value={familySelectId}
-                onChange={(e) => {
-                  const v = e.target.value;
-                  setFamilySelectId(v);
-                  if (v) {
-                    setPendingLocationId(null);
-                    setOptimisticLocationSummary(null);
-                  }
-                }}
-              >
-                <option value=''>None</option>
-                {familyPicker.map((f) => (
-                  <option key={f.id} value={f.id}>
-                    {f.label}
-                  </option>
-                ))}
-              </Select>
-            </div>
-            <div>
-              <Label htmlFor='crm-contact-org'>Organisation</Label>
-              <Select
-                id='crm-contact-org'
-                value={organizationSelectId}
-                onChange={(e) => {
-                  const v = e.target.value;
-                  setOrganizationSelectId(v);
-                  if (v) {
-                    setPendingLocationId(null);
-                    setOptimisticLocationSummary(null);
-                  }
-                }}
-              >
-                <option value=''>None</option>
-                {organizationPicker.map((o) => (
-                  <option key={o.id} value={o.id}>
-                    {o.label}
-                  </option>
-                ))}
-              </Select>
-            </div>
-            {editorMode === 'edit' ? (
-              <div>
-                <Label htmlFor='crm-contact-active'>Status</Label>
-                <Select
-                  id='crm-contact-active'
-                  value={active ? 'true' : 'false'}
-                  onChange={(e) => setActive(e.target.value === 'true')}
-                >
-                  <option value='true'>Active</option>
-                  <option value='false'>Archived</option>
-                </Select>
-              </div>
-            ) : null}
-          </div>
-
-          <AdminCollapsibleSection id='crm-contact-location' title='Location' disabled={isSaving}>
-            <InlineLocationEditor
-              stateKey={inlineLocationStateKey}
-              location={resolvedLocation}
-              embeddedSummary={embeddedLocationSummary}
-              areas={geographicAreas}
-              areasLoading={areasLoading}
-              canModify={!linkedToFamilyOrOrg}
-              readOnlyLockedLines={readOnlyLockedLinesForEditor}
-              readOnlyNote={
-                linkedToFamilyOrOrg
-                  ? 'Location is managed on the linked family or organisation.'
-                  : null
-              }
-              isSaving={isSaving || locationSaveStatus.isSaving}
-              isGeocoding={locationGeocoding}
-              saveError={locationSaveStatus.error}
-              onRequestEdit={() => {}}
-              onCancelEdit={() => {}}
-              onSaveCreate={async (payload) => {
-                const created = await createSharedLocation(payload);
-                if (created) {
-                  setPendingLocationId(created.id);
-                  setOptimisticLocationSummary(summaryFromLocationRow(created));
-                  return created.id;
-                }
-                return null;
-              }}
-              onSaveUpdate={async (id, payload) => {
-                await updateSharedLocation(id, payload);
-              }}
-              onClear={() => {
-                setPendingLocationId(null);
-                setOptimisticLocationSummary(null);
-                clearLocationSaveError();
-              }}
-              onGeocode={geocodeLocation}
-            />
-          </AdminCollapsibleSection>
-
-          <EntityTagPicker
-            id='crm-contact-tags'
-            label='Tags'
-            tags={tags}
-            selectedIds={tagIds}
-            onChange={setTagIds}
-            disabled={isSaving}
-            variant='collapsible'
-          />
-
-          <EntityServicesSection id='crm-contact-services' labels={serviceLabels} />
-
-          {editorMode === 'edit' && selected ? (
-            <div className='text-sm text-slate-600'>
-              <p>Mailchimp: {formatEnumLabel(selected.mailchimp_status)}</p>
-            </div>
-          ) : null}
-        </div>
-      </AdminEditorCard>
-
-      {notesTarget ? (
+      <ContactEditorCard
+        editor={editor}
+        tags={tags}
+        geographicAreas={geographicAreas}
+        areasLoading={areasLoading}
+      />
+      {editor.notesTarget ? (
         <ContactNotesPanel
-          contact={notesTarget}
+          contact={editor.notesTarget}
           adminUsers={adminUsers}
-          onClose={() => setNotesTarget(null)}
+          onClose={() => editor.setNotesTarget(null)}
           onStandaloneNoteCountChange={onPatchStandaloneNoteCount}
         />
       ) : null}
-
-      <PaginatedTableCard
-        title='Contacts'
+      <ContactsListTable
+        rows={rows}
+        filters={filters}
+        setFilter={setFilter}
         isLoading={isLoading}
         isLoadingMore={isLoadingMore}
         hasMore={hasMore}
-        error={error || deleteActionError}
-        loadingLabel='Loading contacts...'
-        onLoadMore={loadMore}
-        toolbar={
-          <AdminTableToolbar>
-            <div className='min-w-[200px] flex-1'>
-              <Label htmlFor='crm-contacts-search'>Search</Label>
-              <Input
-                id='crm-contacts-search'
-                value={filters.query}
-                onChange={(e) => {
-                  setDeleteActionError('');
-                  setFilter('query', e.target.value);
-                }}
-                placeholder='Name, email, phone, Instagram'
-              />
-            </div>
-            <div className='min-w-[140px]'>
-              <Label htmlFor='crm-contacts-type'>Type</Label>
-              <Select
-                id='crm-contacts-type'
-                value={filters.contact_type}
-                onChange={(e) => {
-                  setDeleteActionError('');
-                  setFilter('contact_type', e.target.value as EntityListFilters['contact_type']);
-                }}
-              >
-                <option value=''>All</option>
-                {CONTACT_TYPES.map((v) => (
-                  <option key={v} value={v}>
-                    {formatEnumLabel(v)}
-                  </option>
-                ))}
-              </Select>
-            </div>
-            <div className='min-w-[140px]'>
-              <Label htmlFor='crm-contacts-active'>Status</Label>
-              <Select
-                id='crm-contacts-active'
-                value={filters.active}
-                onChange={(e) => {
-                  setDeleteActionError('');
-                  setFilter('active', e.target.value as EntityListFilters['active']);
-                }}
-              >
-                <option value=''>All</option>
-                <option value='true'>Active</option>
-                <option value='false'>Archived</option>
-              </Select>
-            </div>
-          </AdminTableToolbar>
-        }
-      >
-        <AdminDataTable tableClassName='min-w-[720px]'>
-          <AdminDataTableHead>
-            <tr>
-              <AdminDataTableHeadCell>Name</AdminDataTableHeadCell>
-              <AdminDataTableHeadCell>Email</AdminDataTableHeadCell>
-              <AdminDataTableHeadCell>Type</AdminDataTableHeadCell>
-              <AdminDataTableOperationsHeadCell />
-            </tr>
-          </AdminDataTableHead>
-          <AdminDataTableBody>
-            {rows.map((row) => {
-              const name = [row.first_name, row.last_name].filter(Boolean).join(' ') || '—';
-              const nameListSuffix = contactNameListSuffix(row);
-              return (
-                <tr
-                  key={row.id}
-                  className={`cursor-pointer transition ${
-                    selectedId === row.id ? 'bg-slate-100' : 'hover:bg-slate-50'
-                  }`}
-                  onClick={() => selectRow(row)}
-                >
-                  <AdminDataTableCell>
-                    {name}
-                    {nameListSuffix ? (
-                      <>
-                        <span aria-hidden>{nameListSuffix}</span>
-                        {row.family_ids.length > 0 ? (
-                          <span className='sr-only'>, linked to a family</span>
-                        ) : null}
-                        {row.organization_ids.length > 0 ? (
-                          <span className='sr-only'>, linked to an organisation</span>
-                        ) : null}
-                        {row.relationship_type === 'client' ? (
-                          <span className='sr-only'>, client relationship</span>
-                        ) : null}
-                      </>
-                    ) : null}
-                  </AdminDataTableCell>
-                  <AdminDataTableCell>{row.email ?? '—'}</AdminDataTableCell>
-                  <AdminDataTableCell>{formatEnumLabel(row.contact_type)}</AdminDataTableCell>
-                  <AdminDataTableCell className='text-right'>
-                    <div className='flex flex-wrap justify-end gap-2'>
-                      <Button
-                        type='button'
-                        size='sm'
-                        variant='outline'
-                        className='relative h-8 min-w-8 overflow-visible px-0'
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          setNotesTarget((current) => (current?.id === row.id ? null : row));
-                        }}
-                        disabled={isSaving}
-                        aria-label='Contact notes'
-                        title='Contact notes'
-                      >
-                        <NoteIcon className='h-4 w-4 shrink-0' aria-hidden />
-                        {row.standalone_note_count > 0 ? (
-                          <span className='absolute -right-1 -top-1 flex h-4 min-w-4 items-center justify-center rounded-full bg-rose-600 px-1 text-[10px] font-semibold leading-none text-white'>
-                            {row.standalone_note_count > 99 ? '99+' : row.standalone_note_count}
-                          </span>
-                        ) : null}
-                      </Button>
-                      <Button
-                        type='button'
-                        size='sm'
-                        variant='outline'
-                        className='h-8 min-w-8 px-0'
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          void updateContact(row.id, { active: !row.active }).then(() =>
-                            refreshFamilyOrgLists?.()
-                          );
-                        }}
-                        disabled={isSaving}
-                        aria-label={row.active ? 'Archive contact' : 'Restore contact'}
-                        title={row.active ? 'Archive' : 'Restore'}
-                      >
-                        {row.active ? (
-                          <ArchiveIcon className='h-4 w-4 shrink-0' aria-hidden />
-                        ) : (
-                          <RestoreIcon className='h-4 w-4 shrink-0' aria-hidden />
-                        )}
-                      </Button>
-                      <Button
-                        type='button'
-                        size='sm'
-                        variant='danger'
-                        className='h-8 min-w-8 px-0'
-                        onClick={(e) => {
-                          void handleDeleteContact(row, e);
-                        }}
-                        disabled={isSaving}
-                        aria-label='Delete contact'
-                        title='Delete contact'
-                      >
-                        <DeleteIcon className='h-4 w-4 shrink-0' aria-hidden />
-                      </Button>
-                    </div>
-                  </AdminDataTableCell>
-                </tr>
-              );
-            })}
-          </AdminDataTableBody>
-        </AdminDataTable>
-      </PaginatedTableCard>
+        error={error}
+        loadMore={loadMore}
+        isSaving={editor.isSaving}
+        selectedId={editor.selectedId}
+        deleteActionError={editor.deleteActionError}
+        onClearDeleteError={() => editor.setDeleteActionError('')}
+        onSelectRow={editor.selectRow}
+        onToggleNotes={(row) => {
+          editor.setNotesTarget((current) => (current?.id === row.id ? null : row));
+        }}
+        onToggleActive={(row) => {
+          void editor.updateContact(row.id, { active: !row.active }).then(() =>
+            editor.refreshFamilyOrgLists?.()
+          );
+        }}
+        onDeleteContact={editor.handleDeleteContact}
+      />
     </div>
   );
 }
